Extract DPM wave AI handoff boundary
